rename `rasa.train` [production code]
