+++ tests/cli/test_rasa_train.py
@@ -11,7 +11,7 @@ import rasa.shared.utils.io
 from rasa import model
 from rasa.nlu.model import Metadata
 from rasa.shared.nlu.training_data import training_data
-from rasa.train import (
+from rasa.model_training import (
     CODE_CORE_NEEDS_TO_BE_RETRAINED,
     CODE_NLU_NEEDS_TO_BE_RETRAINED,
     CODE_NLG_NEEDS_TO_BE_RETRAINED,
